add assembly, rename folder devtools in editortools, add namespace
