+++ Assets/Scripts/EditorTools/EditorTools.asmdef
@@ -1,0 +1,3 @@
+{
+	"name": "EditorTools"
+}
