+++ Assets/Scripts/EditorTools/EditorTools.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ff92eadc23e03cb4981819644b0ad66d
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
